+++ src/main/java/com/fullcar/member/application/member/MemberService.java
@@ -5,15 +5,11 @@ import com.fullcar.core.exception.NotFoundException;
 import com.fullcar.core.response.ErrorCode;
 import com.fullcar.member.application.auth.AppleAuthService;
 import com.fullcar.member.application.auth.KakaoAuthService;
-import com.fullcar.member.domain.car.Car;
-import com.fullcar.member.domain.car.CarId;
 import com.fullcar.member.domain.car.CarRepository;
-import com.fullcar.member.domain.mail.Mail;
 import com.fullcar.member.domain.mail.MailRepository;
 import com.fullcar.member.domain.member.*;
 import com.fullcar.member.presentation.member.dto.request.NicknameRequestDto;
 import com.fullcar.member.presentation.member.dto.request.OnboardingRequestDto;
-import com.fullcar.member.presentation.member.dto.request.WithdrawRequestDto;
 import com.fullcar.member.presentation.member.dto.response.MemberGetResponseDto;
 import com.fullcar.member.presentation.member.dto.response.NicknameResponseDto;
 import lombok.RequiredArgsConstructor;
@@ -21,19 +17,12 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.validation.annotation.Validated;
 
-import java.io.IOException;
-import java.util.Objects;
-
 @Validated
 @Service
 @RequiredArgsConstructor
 public class MemberService {
     private final MemberRepository memberRepository;
     private final MemberMapper memberMapper;
-    private final CarRepository carRepository;
-    private final MailRepository mailRepository;
-    private final AppleAuthService appleAuthService;
-    private final KakaoAuthService kakaoAuthService;
 
     /**
      * 회원을 식별자로 조회합니다.
@@ -66,36 +55,5 @@ public class MemberService {
                     .nickname(nicknameRequestDto.getNickname())
                     .build();
         }
-    }
-
-    @Transactional
-    public void withdrawMember(Member member, WithdrawRequestDto withdrawRequestDto) throws IOException {
-        if (Objects.equals(withdrawRequestDto.getSocialType(), "APPLE")) {
-            appleAuthService.revoke(member);
-        }
-        else if (Objects.equals(withdrawRequestDto.getSocialType(), "KAKAO")) {
-            kakaoAuthService.revoke(member);
-        }
-        else {
-            throw new CustomException(ErrorCode.INVALID_MEMBER);
-        }
-
-        deleteCar(member.getCarId());
-        deleteMail(member.getId());
-
-        memberRepository.saveAndFlush(member.deleted());
-
-    }
-
-    private void deleteCar(CarId carId) {
-        if (carRepository.existsByCarId(carId)) {
-            Car car = carRepository.findByCarIdAndIsDeletedOrThrow(carId, false);
-            carRepository.delete(car);
-        }
-    }
-
-    private void deleteMail(MemberId memberId) {
-        Mail mail = mailRepository.findByMemberId(memberId);
-        mailRepository.delete(mail);
     }
 }
