+++ src/main/java/com/fullcar/member/domain/member/service/MailService.java
@@ -1,6 +1,7 @@
 package com.fullcar.member.domain.member.service;
 
 import com.fullcar.member.domain.member.Member;
+import com.fullcar.member.domain.member.MemberId;
 import com.fullcar.member.presentation.member.dto.request.CodeRequestDto;
 import com.fullcar.member.presentation.member.dto.request.EmailRequestDto;
 import org.springframework.transaction.annotation.Transactional;
@@ -10,4 +11,6 @@ public interface MailService {
     void sendMail(Member member, EmailRequestDto emailRequestDto);
 
     void checkMailAuthenticationCode(Member member, CodeRequestDto codeRequestDto);
+
+    void deleteMail(MemberId id);
 }
